Add some flags and clean up config stuff

+++ src/index_runner/main.py
@@ -77,7 +77,8 @@ def main():
     wait_for_dependencies(timeout=180)
     # Used for re-fetching the configuration with a throttle
     last_updated_minute = int(time.time() / 60)
-    _CONFIG_TAG = _fetch_config_tag()
+    if not config()['global_config_url']:
+        config_tag = _fetch_latest_config_tag()
 
     # Database initialization
     es_indexer.init_indexes()
@@ -88,12 +89,13 @@ def main():
         if msg is None:
             continue
         curr_min = int(time.time() / 60)
-        if curr_min > last_updated_minute:
-            config_tag = _fetch_config_tag()
+        if not config()['global_config_url'] and curr_min > last_updated_minute:
             # Check for configuration updates
+            latest_config_tag = _fetch_latest_config_tag()
             last_updated_minute = curr_min
-            if config_tag is not None and config_tag != _CONFIG_TAG:
-                _CONFIG_TAG = config_tag
+            if config_tag is not None and latest_config_tag != config_tag:
+                config(force_reload=True)
+                config_tag = latest_config_tag
                 es_indexer.reload_aliases()
         if msg.error():
             if msg.error().code() == KafkaError._PARTITION_EOF:
@@ -128,7 +130,8 @@ def _handle_msg(msg):
     if event_type in ['REINDEX', 'NEW_VERSION', 'COPY_OBJECT', 'RENAME_OBJECT']:
         obj = _fetch_obj_data(msg)
         ws_info = _fetch_ws_info(msg)
-        releng_importer.run_importer(obj, ws_info, msg)
+        if not config()['skip_releng']:
+            releng_importer.run_importer(obj, ws_info, msg)
         es_indexer.run_indexer(obj, ws_info, msg)
     elif event_type == 'REINDEX_WS' or event_type == 'CLONE_WORKSPACE':
         # Reindex all objects in a workspace, overwriting existing data
@@ -145,7 +148,7 @@ def _handle_msg(msg):
             obj_ref = f"{msg['wsid']}/{msg['objid']}/{msg.get('ver', '?')}"
             obj = _fetch_obj_data(msg)
             ws_info = _fetch_ws_info(msg)
-            if not exists_in_releng:
+            if not exists_in_releng and not config()['skip_releng']:
                 logger.info(f"Importing object {obj_ref} into RE.")
                 releng_importer.run_importer(obj, ws_info, msg)
             if not exists_in_es:
@@ -154,15 +157,18 @@ def _handle_msg(msg):
     elif event_type == 'OBJECT_DELETE_STATE_CHANGE':
         # Delete the object on RE and ES. Synchronous for now.
         es_indexer.delete_obj(msg)
-        releng_importer.delete_obj(msg)
+        if not config()['skip_releng']:
+            releng_importer.delete_obj(msg)
     elif event_type == 'WORKSPACE_DELETE_STATE_CHANGE':
         # Delete everything in RE and ES under this workspace
         es_indexer.delete_ws(msg)
-        releng_importer.delete_ws(msg)
+        if not config()['skip_releng']:
+            releng_importer.delete_ws(msg)
     elif event_type == 'SET_GLOBAL_PERMISSION':
         # Set the `is_public` permissions for a workspace
         es_indexer.set_perms(msg)
-        releng_importer.set_perms(msg)
+        if not config()['skip_releng']:
+            releng_importer.set_perms(msg)
     elif event_type == 'RELOAD_ELASTIC_ALIASES':
         # Reload aliases on ES from the global config file
         es_indexer.reload_aliases()
@@ -211,9 +217,11 @@ def _fetch_ws_info(msg):
     return ws_info
 
 
-def _fetch_config_tag():
-    """using github release api (https://developer.github.com/v3/repos/releases/) find
-    out if there is new version of the config."""
+def _fetch_latest_config_tag():
+    """
+    Using the Github release API, check for a new version of the config.
+    https://developer.github.com/v3/repos/releases/
+    """
     github_release_url = config()['github_release_url']
     if config()['github_token']:
         headers = {'Authorization': f"token {config()['github_token']}"}
